split out tests to seperate files all following similar rules to the original so that we can inject different selectors
